ModelBackEnd/MODELBACKEND/LLM_model/local/ChatAIBackend/voiceguard_backend_skeleton/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
# Optional ASR (uncomment to enable)
# faster-whisper==1.0.0
python-dotenv==1.0.1
